+++ src/app.jsx
@@ -4,32 +4,30 @@ import {
   createBrowserRouter,
   createRoutesFromElements,
 } from "react-router-dom";
-
 import Error from "./components/error";
 import Layout from "./components/layout";
-import Index from "./pages/index";
-import Profile from "./pages/profile";
-
-import userInformation from "./pages/loaders";
-
 import { ButtonEvent } from "./pages/actions";
-import UserInput from "./pages/userInput";
+import Home from "./pages/index";
+import Loader from "./pages/loaders";
+import Profile from "./pages/profile";
+import UserInput from "./pages/add-user";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
     <Route path="/" element={<Layout />} errorElement={<Error />}>
-      <Route index element={<Index />} loader={userInformation} />
+      <Route index element={<Home />} loader={Loader} />
       <Route
         path="users/:id"
         element={<Profile />}
-        loader={userInformation}
+        loader={Loader}
         action={ButtonEvent}
       />
       <Route path="add-edit" element={<UserInput />} action={ButtonEvent} />
       <Route path="*" element={<Error />} />
-    </Route>
-  )
+    </Route>,
+  ),
 );
+
 function App() {
   return <RouterProvider router={router} />;
 }
