+++ src/components/menu-buttons/editButton.jsx
@@ -1,0 +1,11 @@
+import { Link } from "react-router-dom";
+
+function EditButton({ user }) {
+  return (
+    <Link to={"/add-edit"} state={{ user }} className="menuButton">
+      Edit
+    </Link>
+  );
+}
+
+export default EditButton;
