+++ src/components/menu-buttons/index.js
@@ -1,0 +1,3 @@
+export { default as BackButton } from "./BackButton";
+export { default as DeleteButton } from "./DeleteButton";
+export { default as EditButton } from "./EditButton";
