+++ src/hooks/use-organized-list.js
@@ -1,13 +1,13 @@
 import { useState } from "react";
-import { groupContactsByLetterUsingLastNames } from "../../utils";
+import { groupContactsByLetterUsingLastNames } from "../utils";
 
 function useOrganizeList(list) {
   const [search, setSearch] = useState("");
 
   const orderedList = groupContactsByLetterUsingLastNames(
     list.filter((user) =>
-      user.name.toLowerCase().includes(search.toLowerCase())
-    )
+      user.name.toLowerCase().includes(search.toLowerCase()),
+    ),
   );
 
   const keys = Object.keys(orderedList);
