+++ src/pages/index.jsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 import { Await, useLoaderData } from "react-router-dom";
 import Loading from "../components/loading";
-import UserList from "../components/users/userList";
+import UserList from "../components/users";
 
 function Index() {
   const users = useLoaderData();
