+++ vite.config.js
@@ -1,5 +1,5 @@
+import react from "@vitejs/plugin-react-swc";
 import { defineConfig } from "vite";
-import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -17,6 +17,6 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "jsdom",
-    setupFiles: "./src/setupTests.js",
+    setupFiles: "./src/tests/setup-tests.js",
   },
 });
